% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporter-Descriptive.R
\name{docdescriptR}
\alias{docdescriptR}
\title{Descriptive Report in .docx file}
\usage{
docdescriptR(path = "", var.type = "all")
}
\arguments{
\item{path}{(Optional) A character vector with the path to data file. If empty character string (""), interface will appear to choose file.}

\item{var.type}{(Optional) The type of variables to perform analysis, with possible values: "all", "numeric", "integer", "double", "factor", "character".}
}
\value{
The output
  will be a document in the temp folder (tempdir()).
}
\description{
This R Package asks for a .csv file with data and returns a report (.docx) with Descriptive Report concerning all possible variables (i.e. columns).
}
\examples{
\donttest{
data(iris)
write.csv(iris,file="iriscsvfile.csv")
docdescriptR(path="iriscsvfile.csv")
}
}
\references{
\insertAllCited{}
}
